refactor: move concrete channel wrapper into sub module
